chore: clean up and refactor
